feat(trace): Added default values for easier integration.
