+++ example/basic-endpoint.go
@@ -8,11 +8,7 @@ import (
 	"github.com/RisingStack/trace-go/trace"
 )
 
-var collector = trace.NewMemoryCollector()
-
 func handle2(rw http.ResponseWriter, r *http.Request) {
-	span := trace.NewSpanIDFromRequest(r)
-	log.Println("Id in 2nd endpoint: " + span.String() + " on URL: " + r.URL.String())
 	resp := struct {
 		Message string `json:"message"`
 	}{
@@ -27,14 +23,7 @@ func handle(rw http.ResponseWriter, r *http.Request) {
 	}{
 		"Message from Endpoint1 to Endpoint2",
 	}
-	span := trace.NewSpanIDFromRequest(r)
-	log.Println("Id in 1st endpoint: " + span.String())
-	c := http.Client{
-		Transport: trace.Transport{
-			Span:      span,
-			Collector: collector,
-		},
-	}
+	c := http.Client{Transport: trace.NewTransport(r)}
 	_, err := c.Get("http://localhost:9876/test2")
 	if err != nil {
 		log.Fatal(err)
@@ -44,7 +33,7 @@ func handle(rw http.ResponseWriter, r *http.Request) {
 
 func eventsHandler(rw http.ResponseWriter, r *http.Request) {
 	log.Println("Eventshandler called")
-	events := collector.GetEvents()
+	events := trace.DefaultCollector.GetEvents()
 	msg := struct {
 		Events []trace.Event `json:"events"`
 	}{
@@ -54,9 +43,9 @@ func eventsHandler(rw http.ResponseWriter, r *http.Request) {
 }
 
 func main() {
-	http.HandleFunc("/test", trace.Trace(handle, collector))
-	http.HandleFunc("/test2", trace.Trace(handle2, collector))
-	http.HandleFunc("/events", trace.Trace(eventsHandler, collector))
+	http.HandleFunc("/test", trace.Trace(handle))
+	http.HandleFunc("/test2", trace.Trace(handle2))
+	http.HandleFunc("/events", eventsHandler)
 
 	log.Println("Listening on :9876")
 	log.Println(http.ListenAndServe(":9876", nil))
